:truck: [entrypoints] Add cli.create with main function from __main__
